requirements.txt
pyinstaller==6.10.0
maafw==2.1.0
plyer==2.1.0